current_requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.7.2
beautifulsoup4==4.12.3
billiard==4.2.2
celery==5.3.4
certifi==2023.11.17
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==41.0.7
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.20
django-bootstrap5==23.3
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.6
jiter==0.11.0
kombu==5.4.2
lxml==5.3.0
openai==2.2.0
openpyxl==3.1.5
packaging==25.0
pillow==10.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.1
reportlab==4.0.4
requests==2.31.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.4.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.4
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.7.0
